refactor: extract input navigate helpers
